chore(zqlite): move zero cache and zqlite shared filed to a common package
