+++ packages/zero-cache/package.json
@@ -39,7 +39,8 @@
     "postgres-array": "^3.0.2",
     "zero-protocol": "0.0.0",
     "zql": "0.0.0",
-    "xxhash-wasm": "^1.0.2"
+    "xxhash-wasm": "^1.0.2",
+    "zqlite-zero-cache-shared": "0.0.0"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "^0.2.0",
