+++ packages/zqlite-zero-cache-shared/src/lexi-version.test.ts
