+++ packages/zqlite-zero-cache-shared/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {},
+  "include": ["src/**/*.ts", "src/**/*.js"]
+}
