+++ packages/zqlite/package.json
@@ -35,6 +35,7 @@
     "@rocicorp/logger": "^5.2.2",
     "zql": "0.0.0",
     "better-sqlite3": "^11.0.0",
-    "pg-logical-replication": "^2.0.5"
+    "pg-logical-replication": "^2.0.5",
+    "zqlite-zero-cache-shared": "0.0.0"
   }
 }
