--- packages/zqlite/src/duped/readme.md
@@ -1,1 +1,0 @@
-Duped as in duplicated from elsewhere. Original sources to be removed once we've swapped out to the SQLite based replicator.
