--- packages/zqlite/src/replicator/dml-statements.ts
@@ -1,1 +1,0 @@
-export class DmlStatements {}
